update button and delete button works for requests
