+++ jtv-app-frontend/src/components/webpageComponents/homepage.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function homepage() {
+    return (
+        <div>
+            
+        </div>
+    )
+}
